minor changes: rename middleware file to "middlewares", delete extra space, and converted the current validator chains in the /login route into schemas with checkSchema.
