+++ backend/handlers/login.js
@@ -1,6 +1,6 @@
 const validateLogin = async (req, res) => {
   /*
-  Once we establish our data base:
+  Once we establish our database:
     1. query the database for the user based on their username/email
     2. compare password entered to the password stored in the database
     3. then create a session or JWT (not sure which one we are using yet)
